requirements.txt
streamlit
pillow
requests
python-dotenv